requirements.txt
beautifulsoup4==4.5.1
bibtexparser==0.6.2
bs4==0.0.1
Django==1.10
django-autoslug==1.9.3
django-ckeditor==5.1.1
Pillow==3.3.1
psycopg2==2.6.2
requests==2.13.0
